move handle function to endpoint
