+++ packages/api/app/distribution-routes.js
@@ -10,13 +10,15 @@ const { RecordDoesNotExist } = require('@cumulus/errors');
 const { AccessToken } = require('../models');
 
 const {
-  getConfigurations,
+  handleLoginRequest,
+  handleLogoutRequest,
   handleRedirectRequest,
   handleFileRequest,
   useSecureCookies,
+  handleRootRequest,
 } = require('../endpoints/distribution');
 const { isAccessTokenExpired } = require('../lib/token');
-const { buildOAuthClient, checkLoginQueryErrors, getAccessToken, getProfile } = require('../lib/distribution');
+const { getConfigurations, checkLoginQueryErrors, getAccessToken, getProfile } = require('../lib/distribution');
 
 /**
  * Helper function to pull bucket out of a path string.
@@ -98,126 +100,7 @@ async function ensureAuthorizedOrRedirect(req, res, next) {
   return next();
 }
 
-/**
- * Sends a welcome page
- *
- * @param {Object} req - express request object
- * @param {Object} res - express response object
- */
-const root = async (req, res) => {
-  const accessToken = req.cookies.accessToken;
-  let accessTokenRecord;
-  if (accessToken) {
-    try {
-      accessTokenRecord = await new AccessToken().get({ accessToken });
-    } catch (error) {
-      if ((error instanceof RecordDoesNotExist) === false) {
-        throw error;
-      }
-    }
-  }
-
-  const templateVars = {
-    title: 'Welcome',
-    profile: get(accessTokenRecord, 'tokenInfo'),
-    logoutURL: urljoin(process.env.API_BASE_URL, 'logout'),
-  };
-
-  if (!accessToken || !accessTokenRecord) {
-    const authorizeUrl = (await buildOAuthClient()).getAuthorizationUrl(req.path);
-    templateVars.URL = authorizeUrl;
-  }
-
-  const rendered = render(pathresolve(__dirname, 'templates/root.html'), templateVars);
-  return res.send(rendered);
-};
-
 const locate = (req, res) => res.status(501).end();
-
-/**
- * login endpoint
- *
- * @param {Object} req - express request object
- * @param {Object} res - express response object
- * @returns {Promise<Object>} - promise of an express response object
- */
-async function login(req, res) {
-  const { code, state } = req.query;
-  const errorTemplate = pathresolve(__dirname, 'templates/error.html');
-  log.debug('the query params:', req.query);
-  const templateVars = checkLoginQueryErrors(req.query);
-  if (!isEmpty(templateVars) && templateVars.statusCode >= 400) {
-    const rendered = render(errorTemplate, templateVars);
-    return res.type('.html').status(templateVars.statusCode).send(rendered);
-  }
-
-  try {
-    log.debug('pre getAccessToken() with query params:', req.query);
-    const accessTokenResponse = await getAccessToken(code);
-    log.debug('getAccessToken:', accessTokenResponse);
-
-    const userProfile = await getProfile(accessTokenResponse);
-    log.debug('Got the user profile: ', userProfile);
-
-    // expirationTime is in seconds whereas Date is expecting milliseconds
-    const expirationTime = accessTokenResponse.expirationTime * 1000;
-    await new AccessToken().create({
-      accessToken: accessTokenResponse.accessToken,
-      expirationTime,
-      refreshToken: accessTokenResponse.refreshToken,
-      username: accessTokenResponse.username,
-      tokenInfo: userProfile,
-    });
-
-    return res
-      .cookie(
-        'accessToken',
-        accessTokenResponse.accessToken,
-        {
-          expires: new Date(expirationTime),
-          httpOnly: true,
-          secure: useSecureCookies(),
-        }
-      )
-      .status(307)
-      .set({ Location: urljoin(process.env.API_BASE_URL, state || '') })
-      .send('Redirecting');
-  } catch (error) {
-    log.error('Error occurred while trying to login:', error);
-    const vars = {
-      contentstring: `There was a problem talking to OAuth provider, ${error.message}`,
-      title: 'Could Not Login',
-      statusCode: 401,
-    };
-    const rendered = render(errorTemplate, vars);
-    return res.type('.html').status(401).send(rendered);
-  }
-}
-
-/**
- * logout endpoint
- *
- * @param {Object} req - express request object
- * @param {Object} res - express response object
- * @returns {Promise<Object>} - promise of an express response object
- */
-const logout = async (req, res) => {
-  const accessToken = req.cookies.accessToken;
-  const authorizeUrl = (await buildOAuthClient()).getAuthorizationUrl();
-  res.clearCookie('accessToken',
-    {
-      httpOnly: true,
-      secure: useSecureCookies(),
-    });
-  const templateVars = {
-    title: 'Logged Out',
-    contentstring: accessToken ? 'You are logged out.' : 'No active login found.',
-    URL: authorizeUrl,
-    logoutURL: urljoin(process.env.API_BASE_URL, 'logout'),
-  };
-  const rendered = render(pathresolve(__dirname, 'templates/root.html'), templateVars);
-  return res.send(rendered);
-};
 
 const profile = (req, res) => res.send('Profile not available.');
 
@@ -229,10 +112,10 @@ const s3CredentialsREADME = (req, res) => res.status(501).end();
 
 const version = (req, res) => res.status(501).end();
 
-router.get('/', root);
+router.get('/', handleRootRequest);
 router.get('/locate', locate);
-router.get('/login', login);
-router.get('/logout', logout);
+router.get('/login', handleLoginRequest);
+router.get('/logout', handleLogoutRequest);
 router.get('/profile', profile);
 router.get('/pubkey', pubkey);
 router.get('/redirect', handleRedirectRequest);
